add valid syntax cases for send method and add test cases
